+++ ashley/src/test/kotlin/ktx/ashley/TestSystem.kt
@@ -1,0 +1,5 @@
+package ktx.ashley
+
+import com.badlogic.ashley.core.EntitySystem
+
+class TestSystem : EntitySystem()